browser: first cut at horizontal split.
When in 'merge' mode we display the patch in a bottom window when on a
conflict or wiggle, so full context can be see.

This is a bit rough at the moment - the image jumps more than I would
like and it is hard to visually follow what is happening.

Signed-off-by: NeilBrown <neilb@suse.de>