Encourage more context into conflict reports.
Conflict reports currently don't contain much surrounding
'Extraneous' text.  When the patch didn't really apply at all,
this context is very useful.
So encourage at least 3 lines of context from Extraneous or
Unchanged texts.
If it mostly comes from Unchanged text it isn't quite so useful,
so reduce the requirement.

Update self-tests to match.  In most cases these are an improvement.
In a couple of case we maybe get a bit more context than really
necessary, but that is better than not enough.

Signed-off-by: NeilBrown <neilb@suse.de>