Browser: add a more effective cursor.
Create a proper cursor structure to track where on the screen we are.
Record the width of the current symbol so 'backward' works better.
Allow movement master start and end of line by going to previous or
next line.
Allow 'search' to locate the actual string, not just the line.

Signed-off-by: NeilBrown <neilb@suse.de>