merge: make some conflicts less conflicting.
If a 'conflict' has an empty 'before' section, then it is probably
to separate things - and insertion in the original (unmatched) and
and insertion in the result.
So split it into those two, but leave the result marked as a Conflict
as the order of the insertions is not clear.

For now this mostly affects -B which can display the conflict
more effectively.

Signed-off-by: NeilBrown <neilb@suse.de>