vpatch - use 'cols' in main window.
We currently don't use 'cols' - which causes a warning because
it is set by an assignment inside a macro..

The trick to quiet to warning causes different warnings on other
compilers.

We could just do
   rows = getmaxy()
but that is 'legacy' so I'd rather not.

So just find somewhere to use cols..

Reported-by: "P. Martin" <mrsmiley98@lycos.com>
Signed-off-by: NeilBrown <neilb@suse.de>