Hack to improve view of ignore-blank conflicts.
  cd tests/contrib/abstract
  ../../../wiggle -Bb orig new new2
select the "central" and type 'x' so it disappears and note
that the only remaining difference is that "computational"
has been deleted.
Without this patch you only see one '-' line and no '+' line.

However the result isn't perfect as
   ./wiggle -Bp demo.patch
 visit the README file
 page down to waht "You can use 'o'  ..." as added.
And note that there is a '-' blank line and a '+' blank line.
These are unwanted and added by this patch.

Signed-off-by: NeilBrown <neilb@suse.de>