extract: allow blank lines in unified diffs.
When a unified diff report that both files have a blank line,
it shows this as a line containing just a space.
Sometimes that space can go missing (spaces at the end of a line
are like that).
So if we find a completely empty line, treat it like a line
containing just a space.

Resolves-github-issue: #12
Reported-by: luckydonald @ github.com
Signed-off-by: NeilBrown <neilb@suse.de>