diff: filter out unmatchable lines before comparing.
If either file has a run of 2 or more consecutive lines, none of which
appear in the other file, then reduce the run down to a single line.
This does not materially change the set of common-sub-lists that are
calculated, but it can make the calculate much faster when there is a
lot of unmatchable lines.

Signed-off-by: NeilBrown <neil@brown.name>