diff: trim endpoints of search more agressively.
When we find a long snake, we significantly reduce the worst-case
result (which is "no more snakes ever").  This allows us to trim the
end-points of the search-font.  If the endpoints have a best-case that
is worst than the worst-case when using the new snake, there is no point
pursuing them.
Previously we only trimmed the ends one step for each step forward.
This is unnecessarily cautious.  It is better to keep trimming until
the best-case at the end reaches the worst-case.

This improves one test case about 20%.

Signed-off-by: NeilBrown <neil@brown.name>