diff: allocate csl buf incrementally.
The csl buff is currently allocated after the first pass, and filled in
during the second recursively-subdivided pass.
This means the two passes need to produce exactly the same result,
which makes it hard to introduce heuristics to cut corners on
big searches.

So change to allocating incrementally (in powers of 2) as needed.

Signed-off-by: NeilBrown <neil@brown.name>