merge: ensure merge markers are always at start-of-line.
Some unusually circumstances including changes at the end of a file that
doesn't end with "\n" can result in a merge marker (e.g.  <<<<<<<) not
appearing at the start of the line.

This is always bad, so make sure it never happens by keeping track on
whether we are at start of line or not.

Signed-off-by: NeilBrown <neil@brown.name>